Project submission
Reusability:
I designed the app with the "White-label" theme in mind, so I built the components that had brand-specific styling using a Theme context to inject the primary brand color to easily customize the app for each client by only changing one variable.

Added features:
ActionableInfoButton allows the photography vendor to call, get directions, and draft an email. By clicking on the button, the app will bring them to their native maps, phone, and email client to quickly contact their clients.
A drop-down menu in the header allows the user to filter for and get specific types of appointments, which defaults to Scheduled appointments.
The InfoBox component shows the user the most relevant info they need right at the top.
I know this might look weird but:

I used some javascript string manipulation in the AppointmentModal to get the client's Email, Phone, and address because the JSON object had them in a string instead of a properly formatted Object I would usually not take that approach, but to get it working, I used a little "Duck tape."

Added dependencies:
@expo/vector-icons: for vector icons
@react-native-community/datetimepicker: so the user can select dates and times for rescheduling the appointments
expo-linking: so the user can be directed to their respective native app like maps, phone, and email client to execute quick actions
moment: for displaying timestamps in the correct format
react-native-modal: Gives you more features than available with the standard Modal that comes out of the box with react-native, like swipe gestures to make them feel more native and Customizable backdrop opacity, color, etc
expo-clipboard: to copy client info to clipboard in the event we can't launch the native app (Which mostly happens in the simulator but is helpful for that edge case when it doesn't work)
react-native-responsive-screen: for getting the percentage values (wp,hp) for the hight and width of the screen to help make the UI resize according to the device screen dimensions